% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fq.R
\name{add_fq}
\alias{add_fq}
\title{add_fq}
\usage{
add_fq(df, x, start = 10)
}
\arguments{
\item{df}{A data frame}

\item{x}{A date-time variable in \code{df}}

\item{start}{numeric indicating the starting month of a fiscal year}
}
\description{
Add fiscal quarters to a data frame
}
\details{
Creates a new column based on an existing vector of class date-time.
}
\author{
Tyler J. Loewenstein
}
