% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins}
\alias{coins}
\title{coins}
\usage{
coins(x)
}
\arguments{
\item{x}{A double.}
}
\value{
The word "coin" repeated \code{x} times.
}
\description{
Prints the word "coin" a number of times equal to \code{x}
}
\details{
\code{coins} is a test function.
}
\author{
Tyler J. Loewenstein
}
