% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinstarr.R
\docType{package}
\name{coinstarr}
\alias{coinstarr}
\alias{coinstarr-package}
\title{coinstarr: Keeping track of quarters}
\description{
coinstarr aggregates time series data and provides useful tools for finding trends across fiscal quarters.
}
