% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs-data.R
\docType{data}
\name{jobs}
\alias{jobs}
\title{Random data sample of jobs}
\format{An object of class \code{data.frame} with 10000 rows and 8 columns.}
\usage{
data(jobs)
}
\description{
Data is a random sample of fake employment positions
}
\examples{
data(jobs)
vacancies <- collect(jobs, "agency", method = "position")
}
\keyword{datasets}
