% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_by.R
\name{join_by}
\alias{join_by}
\title{join_by}
\usage{
join_by(dfx, dfy, dfz, by = c("agency", "payday"))
}
\arguments{
\item{dfx}{A data frame (likely `jobs`).}

\item{dfy}{A data frame (likely `overtime`).}

\item{dfz}{A data frame (likely `leave`).}

\item{by}{One or more grouping variables, typically `agency` and `payday`.}
}
\description{
Creates a single data frame by combining \code{jobs}, \code{overtime},
and \code{leave} according to the `by` argument.
}
\author{
Tyler J. Loewenstein
}
