% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave-data.R
\docType{data}
\name{leave}
\alias{leave}
\title{Random data sample of leave entries}
\format{An object of class \code{data.frame} with 10000 rows and 10 columns.}
\usage{
data(leave)
}
\description{
Data is a random sample of fake leave entries
}
\examples{
data(leave)
vacancies <- collect(leave, "agency", method = "leave")
}
\keyword{datasets}
