% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{merge}
\usage{
merge(dfx, dfy, dfz)
}
\arguments{
\item{dfx}{A data frame (likely `jobs`).}

\item{dfy}{A data frame (likely `overtime`).}

\item{dfz}{A data frame (likely `leave`).}
}
\description{
Merges creates a single data frame by combining \code{jobs}, \code{overtime},
and \code{leave}.
}
\author{
Tyler J. Loewenstein
}
