% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{model}
\usage{
model(df, formula, agency = NULL)
}
\arguments{
\item{df}{A data frame}

\item{formula}{An object of class "formula"}

\item{agency}{If `NULL`, all agencies are included in the regression. If set
to "each", a model for each agency will be produced. Otherwise, argument may
be set to filter a specific agency.}
}
\description{
Generates a linear model for the variables of interest and given agency
}
\details{
A useful adaption to \code{lm()} with an agency filter.
}
\author{
Tyler J. Loewenstein
}
