% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overtime-data.R
\docType{data}
\name{overtime}
\alias{overtime}
\title{Random data sample of overtime entries}
\format{An object of class \code{data.frame} with 10000 rows and 10 columns.}
\usage{
data(overtime)
}
\description{
Data is a random sample of fake overtime entries
}
\examples{
data(overtime)
vacancies <- collect(overtime, "agency", method = "overtime")
}
\keyword{datasets}
