% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{collect}
\usage{
collect(df, group1, group2, method = "position")
}
\arguments{
\item{df}{a data frame}

\item{group1}{a grouping variable}

\item{group2}{a second grouping variable different from group1}

\item{method}{a method}
}
\description{
Summarizes raw agency data into vacancies, overtime hours, and leave hours
}
\details{
This is where the details will go.
}
\author{
Tyler J. Loewenstein
}
