% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{model}
\usage{
model(df, formula, filter = NULL)
}
\arguments{
\item{df}{A data frame}

\item{formula}{An object of class "formula"}

\item{filter}{A logical test to filter observations in the data frame}
}
\description{
Generates a linear model for the variables of interest and given filter.
}
\details{
A useful adaption to \code{lm()} with an integrated filter
}
\author{
Tyler J. Loewenstein
}
